+++ src/main/java/com/wow/server/data/model/Product.java
@@ -18,36 +18,37 @@ import javax.persistence.Table;
 @Setter
 public class Product {
 
-	@Id @GeneratedValue(strategy = GenerationType.AUTO)
-	Long productId;
-	
-	String productCode;
-	
-	String productName;
-	
-	BigDecimal marketCap;
-	
-	BigDecimal avgVolume;
-	
-	BigDecimal peRatio;
-	
-	BigDecimal revenue;
-	
-	BigDecimal totalCash;
-	
-	BigDecimal totalDevidendYield;
-	
-	BigDecimal avgDividendYield;
-	
-	String sector;
-	
-	String industry;
-	
-	BigDecimal price;
-	
-	BigDecimal priceOneDayChange;
-	
-	LocalDateTime creationDateTime;
-	
-	LocalDateTime updateDateTime;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    Long productId;
+
+    String productCode;
+
+    String productName;
+
+    BigDecimal marketCap;
+
+    BigDecimal avgVolume;
+
+    BigDecimal peRatio;
+
+    BigDecimal revenue;
+
+    BigDecimal totalCash;
+
+    BigDecimal totalDevidendYield;
+
+    BigDecimal avgDividendYield;
+
+    String sector;
+
+    String industry;
+
+    BigDecimal price;
+
+    BigDecimal priceOneDayChange;
+
+    LocalDateTime creationDateTime;
+
+    LocalDateTime updateDateTime;
 }
