+++ src/main/java/com/wow/server/data/model/User.java
@@ -5,12 +5,9 @@ import lombok.Setter;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name = "TB_USER")
@@ -18,37 +15,35 @@ import javax.persistence.Table;
 @Setter
 public class User {
 
-	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-	Long userId;
-	
-	String userName;
-	
-	String email;
-	
-	String password;
-	
-	String displayName;
-	
-	LocalDate firstTradeDate;
-	
-	Integer riskProfile;
-	
-	String assetSizeRange;
-	
-	LocalDate dateOfBirth;
-	
-	String educationLevel;
-	
-	String haveChild;
-	
-	String maritalStatus;
-	
-	String monthlyIncomeRange;
-	
-	String introduction;
-	
-	LocalDateTime creationDateTime;
-	
-	LocalDateTime updateDateTime;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long userId;
+    private String userName;
+    private String email;
+    private String password;
+    private String displayName;
+    private LocalDate firstTradeDate;
+    private Integer riskProfile;
+    private String assetSizeRange;
+    private LocalDate dateOfBirth;
+    private String educationLevel;
+    private String haveChild;
+    private String maritalStatus;
+    private String monthlyIncomeRange;
+    private String introduction;
+    private LocalDateTime creationDateTime;
+    private LocalDateTime updateDateTime;
+
+    @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    @JoinTable(
+            name = "TB_FOLLOW",
+            joinColumns = @JoinColumn(name = "userId"),
+            inverseJoinColumns = @JoinColumn(name = "followerId")
+    )
+    public List<User> following;
+
+
+    @ManyToMany(mappedBy = "following", fetch = FetchType.LAZY)
+    public List<User> followers;
 
 }
