+++ src/main/java/com/wow/server/data/repository/UserRepository.java
@@ -4,6 +4,7 @@ import com.wow.server.data.model.User;
 import com.wow.server.data.projection.UserMinimalProjection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
