+++ src/main/java/com/wow/server/user/following/FollowRepository.java
@@ -1,0 +1,9 @@
+package com.wow.server.user.following;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface FollowRepository extends JpaRepository<FollowEntity, Long>{
+
+}
