+++ src/main/java/com/wow/server/user/following/FollowerInformationDTO.java
@@ -1,0 +1,9 @@
+package com.wow.server.user.following;
+
+import lombok.Data;
+
+@Data
+public class FollowerInformationDTO {
+    private Long userId;
+    private String displayName;
+}
