+++ src/main/resources/data.sql
@@ -33,8 +33,8 @@ INSERT INTO TB_FOLLOW
 (
 FOLLOW_ID,
 CREATION_DATE_TIME,
-FOLLOWER_USER_ID,
-FOLLOWING_USER_ID,
+USER_ID,
+FOLLOWER_ID,
 UPDATE_DATE_TIME
 )
 VALUES
