--- src/test/java/com/wow/server/controller/ProductControllerTest.java
@@ -1,74 +1,0 @@
-package com.wow.server.controller;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.google.common.collect.Sets;
-import com.wow.server.DataPreparationUtils;
-import com.wow.server.data.repository.ProductRepository;
-import com.wow.server.mapper.HoldingMapperImpl;
-import com.wow.server.mapper.ProductMapperImpl;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import static org.hamcrest.Matchers.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@AutoConfigureMockMvc
-@ContextConfiguration(classes = {ProductController.class, ProductMapperImpl.class})
-@WebMvcTest
-@ActiveProfiles("test")
-public class ProductControllerTest {
-
-    private static ObjectMapper objectMapper;
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @MockBean
-    private ProductRepository productRepository;
-
-    @BeforeAll
-    public static void initOnce() {
-        objectMapper = Jackson2ObjectMapperBuilder
-                .json()
-                .featuresToDisable(SerializationFeature.WRAP_ROOT_VALUE)
-                .build();
-    }
-
-    @Test
-    public void endpoint_can_return_data_successfully() throws Exception {
-        // given
-        Set<Long> productIds = Sets.newHashSet(34567L,34568L);
-        DataPreparationUtils.mockProductRepositoryResponse(productIds, productRepository);
-
-        // when
-        ResultActions result = mockMvc.perform(post("/api/products")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsBytes(productIds))
-                .accept(MediaType.APPLICATION_JSON));
-
-        // then
-        result
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$", notNullValue()))
-                .andExpect(jsonPath("$", hasSize(2)))
-                .andExpect(jsonPath("$[*].productId", everyItem(in(productIds.stream()
-                        .map(Long::intValue)
-                        .collect(Collectors.toList())))));
-    }
-}
